Prevent flicker in select when projects load
The project name flickers and sometimes the create project form appears
briefly, then disappears, as the projects are loaded in the
project-select component.  Remove the placeholder text to avoid some
flicker. (It should never actually be displayed unless you have no
projects and can't create one, in which case you won't be able to
continue in the wizard anyway.) Also don't show the create project form
if the projects haven't loaded yet.