Allow for a 'file' type in the openshift_form_defintion.
Specifying this type in the openshift_form_defintion will trigger
the necessary schema & form setup so the parameter is shown as a
file upload input in the catalog provision/bind wizard.

The angular-schema-form add-on at https://github.com/Textalk/angular-schema-form-base64-file-upload
is used for the display of the file upload input.
The file uploaded is converted to base64 and stored as a string
parameter upon provsion/bind.