Add limited 'oneOf' json schema dependencies support for plan schemas.
This change examinse the json schema object in the service plan,
looking for a `dependencies` object.
If this exists, any `oneOf` property dependencies defined will be
examined and a corresponding angular-schema-form `condition` will
be added to the form definition so specified fields will only be
shown when the corresponding option is selected/entered in another
field.

Example service plan that has a `oneOf` dependency for changing the
follow up fields depending on the selected value in the first selection
field.
https://gist.github.com/david-martin/594464b8193fb3f81c9bd878709731b5